text input and result

+++ src/main/java/uow/itpm/teamblue/model/repo/DocumentRepository.java
@@ -3,6 +3,4 @@ package uow.itpm.teamblue.model.repo;
 import org.springframework.data.repository.CrudRepository;
 import uow.itpm.teamblue.model.Document;
 
-public interface DocumentRepository extends CrudRepository<Document, Long> {
-
-}
+public interface DocumentRepository extends CrudRepository<Document, Integer> {}
